+++ packages/ads/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @happyvertical/smrt-ads
 
+## 0.21.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.21.39
+  - @happyvertical/smrt-assets@0.21.39
+  - @happyvertical/smrt-commerce@0.21.39
+  - @happyvertical/smrt-properties@0.21.39
+  - @happyvertical/smrt-tags@0.21.39
+  - @happyvertical/smrt-tenancy@0.21.39
+
+## 0.21.38
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.21.38
+  - @happyvertical/smrt-assets@0.21.38
+  - @happyvertical/smrt-commerce@0.21.38
+  - @happyvertical/smrt-properties@0.21.38
+  - @happyvertical/smrt-tags@0.21.38
+  - @happyvertical/smrt-tenancy@0.21.38
+
+## 0.21.37
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.21.37
+  - @happyvertical/smrt-assets@0.21.37
+  - @happyvertical/smrt-commerce@0.21.37
+  - @happyvertical/smrt-properties@0.21.37
+  - @happyvertical/smrt-tags@0.21.37
+  - @happyvertical/smrt-tenancy@0.21.37
+
 ## 0.21.36
 
 ### Patch Changes