+++ packages/smrt-playground/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @happyvertical/smrt-playground
 
+## 0.21.39
+
+### Patch Changes
+
+- @happyvertical/smrt-svelte@0.21.39
+
+## 0.21.38
+
+### Patch Changes
+
+- @happyvertical/smrt-svelte@0.21.38
+
+## 0.21.37
+
+### Patch Changes
+
+- @happyvertical/smrt-svelte@0.21.37
+
 ## 0.21.36
 
 ### Patch Changes