+++ packages/smrt-svelte/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @happyvertical/smrt-svelte
 
+## 0.21.39
+
+### Patch Changes
+
+- @happyvertical/smrt-agents@0.21.39
+- @happyvertical/smrt-jobs@0.21.39
+- @happyvertical/smrt-profiles@0.21.39
+- @happyvertical/smrt-users@0.21.39
+- @happyvertical/smrt-types@0.21.39
+
+## 0.21.38
+
+### Patch Changes
+
+- @happyvertical/smrt-agents@0.21.38
+- @happyvertical/smrt-jobs@0.21.38
+- @happyvertical/smrt-profiles@0.21.38
+- @happyvertical/smrt-users@0.21.38
+- @happyvertical/smrt-types@0.21.38
+
+## 0.21.37
+
+### Patch Changes
+
+- @happyvertical/smrt-agents@0.21.37
+- @happyvertical/smrt-jobs@0.21.37
+- @happyvertical/smrt-profiles@0.21.37
+- @happyvertical/smrt-users@0.21.37
+- @happyvertical/smrt-types@0.21.37
+
 ## 0.21.36
 
 ### Patch Changes