+++ packages/tenancy/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @happyvertical/smrt-tenancy
 
+## 0.21.39
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.21.39
+  - @happyvertical/smrt-users@0.21.39
+  - @happyvertical/smrt-types@0.21.39
+
+## 0.21.38
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.21.38
+  - @happyvertical/smrt-users@0.21.38
+  - @happyvertical/smrt-types@0.21.38
+
+## 0.21.37
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.21.37
+  - @happyvertical/smrt-users@0.21.37
+  - @happyvertical/smrt-types@0.21.37
+
 ## 0.21.36
 
 ### Patch Changes